+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ### v1.2 (unreleased)
 #### Changed
 - refuses remote connection request when it is already under remote control by a different host
+- LED controller waits for acknowledgement from Arduino that it has cleared the display
+#### Fixed
+- handle Arduino microsecond timer rollover during synchronization
 
 ---